src/main/webapp/app/entities/my-service/delete/my-service-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { MyServiceService } from '../service/my-service.service';

import { MyServiceDeleteDialogComponent } from './my-service-delete-dialog.component';

describe('MyService Management Delete Component', () => {
  let comp: MyServiceDeleteDialogComponent;
  let fixture: ComponentFixture<MyServiceDeleteDialogComponent>;
  let service: MyServiceService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, MyServiceDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(MyServiceDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(MyServiceDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(MyServiceService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
